+++ section7/springsecsection7/src/main/java/com/eazybytes/repository/LoanRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.eazybytes.model.Loans;
 
 @Repository
-public interface LoanRepository extends CrudRepository<Loans, Long> {
+public interface LoanRepository extends CrudRepository<Loans, Integer> {
 	
 	List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
