+++ section7/springsecsection7/src/main/java/com/eazybytes/repository/NoticeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.eazybytes.model.Notice;
 
 @Repository
-public interface NoticeRepository extends CrudRepository<Notice, Long> {
+public interface NoticeRepository extends CrudRepository<Notice, Integer> {
 	
 	@Query(value = "from Notice n where CURDATE() BETWEEN noticBegDt AND noticEndDt")
 	List<Notice> findAllActiveNotices();
