+++ section8/springsecsection8/src/main/java/com/eazybytes/repository/CustomerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CustomerRepository extends CrudRepository<Customer, Long> {
+public interface CustomerRepository extends CrudRepository<Customer, Integer> {
 
     List<Customer> findByEmail(String email);
